Linked nodemailer to handlebars, added company email and pw to dotenv
