+++ .vscode/settings.json
@@ -1,3 +1,12 @@
 {
-  "cSpell.words": ["Briscoe", "exphbs", "JAWSDB", "resave", "sess", "signup"]
+  "cSpell.words": [
+    "Briscoe",
+    "cachefood",
+    "exphbs",
+    "JAWSDB",
+    "ngmeeyczbxccfaip",
+    "resave",
+    "sess",
+    "signup"
+  ]
 }
