+++ controllers/api/userRoutes.js
@@ -12,8 +12,8 @@ router.post("/signup", async (req, res) => {
     });
 
     req.session.save(() => {
-      // req.session.user_id = userData.id;
-      // req.session.email = userData.email;
+      req.session.user_id = userData.id;
+      req.session.email = userData.email;
       req.session.logged_in = true;
 
       res.status(200).json({ user: userData, message:"New user created. You are now logged in." });
