+++ package.json
@@ -33,6 +33,7 @@
     "express-session": "^1.17.3",
     "mysql2": "^2.3.3",
     "nodemailer": "^6.7.7",
+    "nodemailer-express-handlebars": "^5.0.0",
     "sequelize": "^6.21.3"
   }
 }
