--- utils/sandbox.html
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-</head>
-
-<body>
-    <p>"The recipe Soy-and-Ginger-Glazed Salmon with Udon Noodles can be made <b>in approximately 1 hour and 35 minutes
-        </b>.
-        One portion of this dish contains about <b>48g of protein </b>, <b>17g of fat </b>, and a total of <b>552
-            calories </b>.
-        This recipe serves 4. For <b>$5.91 per serving </b>, this recipe <b>covers 47% </b> of your daily requirements
-        of
-        vitamins and minerals. It works well as a main course. 1 person has tried and liked this recipe. It is brought
-        to you by
-        Food and Wine. If you have fresh ginger, udon noodles, salmon fillets, and a few other ingredients on hand, you
-        can make
-        it. It is a good option if you're following a <b>dairy free and pescatarian </b> diet. All things considered, we
-        decided
-        this recipe <b>deserves a spoonacular score of 92% </b>. This score is great. If you like this recipe, take a
-        look at
-        these similar recipes: Salmon With Soy-ginger Noodles, Ginger-Soy Salmon With Soba Noodles, and Soy & ginger
-        salmon with
-        soba noodles.",</p>
-</body>
-
-</html>
